Move LLM calling utilities to `mlflow/genai/utils/llm_utils.py`
Extract `_call_llm`, `_TokenCounter`, `_pydantic_to_response_format`,
`_ModelCost`, and `_lookup_model_cost` from discovery/utils into a
shared module so they can be reused across genai domains. Re-export
from discovery/utils for backwards compatibility.
